Added grunt configurations to allow for working in the spotify folder OR a separate folder
The default task runs jshint and jasmine followed by a watch task that will re-run jshint and jasmine whenever a project js file changes.
The deploy task does the same but includes a copy task after jshint and jasmine. The copy task copies the project files to the spotify folder so it can be run.

Use the former if you clone the repository to the spotify folder and do all your work in there. The latter is useful if you prefer to work in a different folder.