A workaround to stop the reflections disappearing
Reflecting the whole li seems to be a bit flaky and I don't understand why - which is bad.
(Some of the reflections disappear at the end of a transition!?)

Just reflecting the img works fine as a workaround.

It means the play button doesn't get reflected (if there is one) but I think that's preferable to that horrible disappearing reflections thing that was happening.